web/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
babel==2.8.0              # via prices
boto3==1.11.5             # via django-amazon-ses
botocore==1.14.5          # via boto3, s3transfer
cachetools==4.1.0         # via google-auth
certifi==2019.11.28       # via requests, sentry-sdk
chardet==3.0.4            # via requests
defusedxml==0.6.0         # via python3-openid
django-allauth==0.41.0    # via -r requirements.in
django-amazon-ses==3.0.1  # via -r requirements.in
django-bootstrap-form==3.4  # via -r requirements.in
django==2.2.9             # via -r requirements.in, django-allauth, django-amazon-ses, django-bootstrap-form
djangorestframework==3.9.4  # via -r requirements.in
docutils==0.15.2          # via botocore
gevent==1.4.0             # via gunicorn
google-auth==1.16.0       # via -r requirements.in
google-measurement-protocol==1.0.0  # via -r requirements.in
greenlet==0.4.15          # via gevent
gunicorn[gevent]==20.0.4  # via -r requirements.in
idna==2.8                 # via requests
jmespath==0.9.4           # via boto3, botocore
newrelic==5.16.2.147      # via -r requirements.in
oauthlib==3.1.0           # via requests-oauthlib
prices==1.0.0             # via google-measurement-protocol
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
python-dateutil==2.8.1    # via botocore
python3-openid==3.1.0     # via django-allauth
pytz==2019.3              # via babel, django
requests-oauthlib==1.3.0  # via django-allauth
requests==2.22.0          # via django-allauth, google-measurement-protocol, requests-oauthlib
rsa==4.0                  # via google-auth
s3transfer==0.3.1         # via boto3
sentry-sdk==0.14.0        # via -r requirements.in
six==1.14.0               # via google-auth, python-dateutil
sqlparse==0.3.0           # via django
typing==3.7.4.1           # via google-measurement-protocol, prices
urllib3==1.25.7           # via botocore, requests, sentry-sdk

# The following packages are considered to be unsafe in a requirements file:
# setuptools
